test: consolidate tests, add xdist + timeout, split CI py matrix by event
L1 — process model
- add pytest-xdist + pytest-timeout to dev deps
- pytest.ini: global 30 s wall-clock, thread-based to coexist with xdist
- tests/conftest.py: per-marker timeout boosts (opensandbox/openviking 300 s,
  integration/bench 600 s, live_llm 180 s) so slow paths inherit headroom
  without touching individual files

L2 — CI matrix layered by event
- .github/workflows/ci-test-fast.yml: split into test-fast-pr (py 3.10 + 3.13)
  and test-fast-main (full 3.10/3.11/3.12/3.13 on push); both add `-n auto`.
  Halves PR minutes; push-to-main still tests the middle versions.

L3 — file consolidation (every test function name preserved, 0 case dropped)
- tests/session/ 17 → 10:
    test_session_core.py     (primitives + Session.create + registry)
    test_lineage.py          (graph unit + JSONL export)
    test_chunk_wire.py       (chunk_table_to_messages + to_create_kwargs)
    test_loop_local.py       (loop happy/error/cap/growth/streaming, all 4 files)
- tests/memory/backends/ openviking 7 → 3:
    test_openviking_canary.py (skipgate + metadata)
    test_openviking_io.py     (lifecycle + fs_ops + mutations)
    test_openviking_query.py  (search + errors)
- tests/session/persistence/ 8 → 4:
    test_loader_gc.py         (loader + gc)
    test_storage_meta.py      (paths + migrations + write_lock + resume_reattach)
    test_writer.py now also carries the 3 loop_persist cases
- tests/backends/persistence/ 3 → 2: test_registry_gc.py (registry + gc)
- tests/memory/persistence/ 2 → 1: test_persistence.py (paths + registry)

Totals: tests/test_*.py 150 → 99 (-34%); collected cases 799 → 799 (守恒).

Validation (local; opensandbox-server + openviking-server in Docker):
- ruff check + format check + mypy: clean
- fast (`-n auto`, with both extras installed): 728 passed, 7 skipped
- opensandbox: 42 passed, 3 skipped
- openviking: 32 passed

Why: cuts maintenance surface and CI wall-clock without changing semantics —
real-only backend tests untouched, multi-Python coverage preserved on main.
