+++ pyproject.toml
@@ -59,6 +59,8 @@ dev = [
     "pytest>=9.0.3",
     "pytest-benchmark>=4.0.0",
     "pytest-rerunfailures>=15.0",
+    "pytest-timeout>=2.3.0",
+    "pytest-xdist>=3.6.0",
     "ruff>=0.15.12",
 ]
 docs = [
