+++ pytest.ini
@@ -1,8 +1,15 @@
 [pytest]
 # Markers (--strict-markers-friendly). Quick local run:
-#   uv run pytest -m "not opensandbox and not live_llm"
+#   uv run pytest -n auto -m "not opensandbox and not openviking and not live_llm"
 # Skip full stack only:
 #   uv run pytest -m "not integration"
+#
+# Per-test wall-clock budget so a hung socket/subprocess can't stall the suite.
+# Integration / bench / live-backend tests bump the budget via
+# `@pytest.mark.timeout(N)`. ``timeout_method = thread`` because the default
+# ``signal`` method only works on the main thread, which fights xdist.
+timeout = 30
+timeout_method = thread
 markers =
     opensandbox: requires local opensandbox-server / Docker (slow)
     opensandbox_real: alias for opensandbox — real opensandbox-server only, no mocks
