+++ tests/session/persistence/test_loader_gc.py
@@ -1,7 +1,15 @@
-"""Round-trip + error-path tests for :func:`load_session` / :func:`list_persisted_sessions`."""
+"""Session persistence: ``load_session`` round-trip + GC helpers.
+
+Consolidated from (every test function name preserved verbatim):
+- test_loader.py    (load_session / list_persisted_sessions / to_resumable_pair)
+- test_gc.py        (delete_session / prune_sessions)
+"""
 
 from __future__ import annotations
 
+import json
+import uuid
+from datetime import datetime, timedelta, timezone
 from pathlib import Path
 from uuid import uuid4
 
@@ -20,8 +28,10 @@ from rath.session.graph.kind import LineageKind
 from rath.session.persistence import (
     PersistenceError,
     SessionWriter,
+    delete_session,
     list_persisted_sessions,
     load_session,
+    prune_sessions,
     session_file,
 )
 from rath.session.session import Session
@@ -38,6 +48,18 @@ def _build_full_session() -> Session:
         lineage_extras=(("loop.truncated", True), ("note", "demo")),
     )
     return s
+
+
+def _write_session(content: str = "hi") -> Session:
+    s = Session(chunk_table=ChunkTable(rows=()))
+    with SessionWriter(s) as w:
+        w.write_chunk(0, user_text_chunk(content))
+    return s
+
+
+# ---------------------------------------------------------------------------
+# load_session round-trip + error paths
+# ---------------------------------------------------------------------------
 
 
 def test_round_trip_full_session(_isolate_openrath_home: Path) -> None:
@@ -126,8 +148,7 @@ def test_list_persisted_sessions_orders_by_created_at(
         w.write_chunk(0, user_text_chunk("one"))
     # The next session's header carries a later created_at since datetime.now
     # increases between calls; if the test environment is too fast we
-    # explicitly nudge via a no-op sleep. The clock resolution we observed in
-    # CI is fine without one, but make the intent explicit:
+    # explicitly nudge via a no-op sleep.
     import time
 
     time.sleep(0.005)
@@ -190,3 +211,87 @@ def test_session_file_path_under_resolved_dir(_isolate_openrath_home: Path) -> N
     sid = uuid4()
     expected = session_file(sid)
     assert str(expected).endswith(f"{sid}.jsonl")
+
+
+# ---------------------------------------------------------------------------
+# GC: delete + prune
+# ---------------------------------------------------------------------------
+
+
+def test_delete_session_removes_file(_isolate_openrath_home: Path) -> None:
+    s = _write_session()
+    assert session_file(s.id).is_file()
+    assert delete_session(s.id) is True
+    assert not session_file(s.id).is_file()
+
+
+def test_delete_session_returns_false_when_absent(
+    _isolate_openrath_home: Path,
+) -> None:
+    assert delete_session(uuid.uuid4()) is False
+
+
+def test_delete_session_is_idempotent(_isolate_openrath_home: Path) -> None:
+    s = _write_session()
+    assert delete_session(s.id) is True
+    assert delete_session(s.id) is False
+
+
+def test_prune_sessions_removes_only_older_than_cutoff(
+    _isolate_openrath_home: Path,
+) -> None:
+    old = _write_session("old one")
+    new = _write_session("new one")
+
+    # Rewrite the old session's header timestamp to long ago by editing the
+    # file directly — this is the cleanest way to simulate "older than".
+    old_path = session_file(old.id)
+    lines = old_path.read_text(encoding="utf-8").splitlines()
+    header = json.loads(lines[0])
+    ancient = datetime.now(timezone.utc) - timedelta(days=400)
+    header["created_at"] = ancient.isoformat()
+    lines[0] = json.dumps(header, ensure_ascii=False)
+    old_path.write_text("\n".join(lines) + "\n", encoding="utf-8")
+
+    removed = prune_sessions(older_than=timedelta(days=30))
+    assert removed == [old.id]
+    surviving = {m.id for m in list_persisted_sessions()}
+    assert surviving == {new.id}
+
+
+def test_prune_sessions_empty_when_nothing_old(_isolate_openrath_home: Path) -> None:
+    _write_session("fresh")
+    assert prune_sessions(older_than=timedelta(days=30)) == []
+
+
+def test_prune_sessions_skips_unparseable_files(
+    _isolate_openrath_home: Path,
+) -> None:
+    """Files that don't parse aren't pruned (manual cleanup is safer)."""
+    from rath.session.persistence.paths import ensure_sessions_dir
+
+    target_dir = ensure_sessions_dir()
+    junk = target_dir / "not-a-uuid.jsonl"
+    junk.write_text("garbage\n", encoding="utf-8")
+    # Should not raise; just returns empty.
+    removed = prune_sessions(older_than=timedelta(seconds=0))
+    assert removed == []
+    assert junk.is_file()
+
+
+@pytest.mark.parametrize("days", [1, 7, 365])
+def test_prune_sessions_respects_arbitrary_cutoffs(
+    _isolate_openrath_home: Path, days: int
+) -> None:
+    s = _write_session("s")
+    path = session_file(s.id)
+    lines = path.read_text(encoding="utf-8").splitlines()
+    header = json.loads(lines[0])
+    header["created_at"] = (
+        datetime.now(timezone.utc) - timedelta(days=days + 1)
+    ).isoformat()
+    lines[0] = json.dumps(header, ensure_ascii=False)
+    path.write_text("\n".join(lines) + "\n", encoding="utf-8")
+
+    removed = prune_sessions(older_than=timedelta(days=days))
+    assert removed == [s.id]
